+++ docs/html/search/all_12.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['untildatetime_290',['untilDateTime',['../structace__time_1_1extended_1_1MatchingEraTemplate.html#a81741d11d0848d186ee235a3b0654b0e',1,'ace_time::extended::MatchingEraTemplate::untilDateTime()'],['../structace__time_1_1extended_1_1TransitionTemplate.html#a526439328e2769dc07b3293d368b555e',1,'ace_time::extended::TransitionTemplate::untilDateTime()']]],
-  ['untilyear_291',['untilYear',['../structace__time_1_1internal_1_1ZoneContext.html#a766f3740d801f4cad83cd955b7489a02',1,'ace_time::internal::ZoneContext']]]
+  ['untildatetime_289',['untilDateTime',['../structace__time_1_1extended_1_1MatchingEraTemplate.html#a81741d11d0848d186ee235a3b0654b0e',1,'ace_time::extended::MatchingEraTemplate::untilDateTime()'],['../structace__time_1_1extended_1_1TransitionTemplate.html#a526439328e2769dc07b3293d368b555e',1,'ace_time::extended::TransitionTemplate::untilDateTime()']]],
+  ['untilyear_290',['untilYear',['../structace__time_1_1internal_1_1ZoneContext.html#a766f3740d801f4cad83cd955b7489a02',1,'ace_time::internal::ZoneContext']]]
 ];