+++ docs/html/search/classes_0.js
@@ -1,11 +1,11 @@
 var searchData=
 [
-  ['basiczone_319',['BasicZone',['../classace__time_1_1BasicZone.html',1,'ace_time']]],
-  ['basiczonemanager_320',['BasicZoneManager',['../classace__time_1_1BasicZoneManager.html',1,'ace_time']]],
-  ['basiczoneprocessor_321',['BasicZoneProcessor',['../classace__time_1_1BasicZoneProcessor.html',1,'ace_time']]],
-  ['basiczoneprocessorcache_322',['BasicZoneProcessorCache',['../classace__time_1_1BasicZoneProcessorCache.html',1,'ace_time']]],
-  ['basiczoneprocessortemplate_323',['BasicZoneProcessorTemplate',['../classace__time_1_1BasicZoneProcessorTemplate.html',1,'ace_time']]],
-  ['basiczoneprocessortemplate_3c_20basic_3a_3abrokerfactory_2c_20basic_3a_3azoneinfobroker_2c_20basic_3a_3azoneerabroker_2c_20basic_3a_3azonepolicybroker_2c_20basic_3a_3azonerulebroker_20_3e_324',['BasicZoneProcessorTemplate&lt; basic::BrokerFactory, basic::ZoneInfoBroker, basic::ZoneEraBroker, basic::ZonePolicyBroker, basic::ZoneRuleBroker &gt;',['../classace__time_1_1BasicZoneProcessorTemplate.html',1,'ace_time']]],
-  ['brokerfactory_325',['BrokerFactory',['../classace__time_1_1internal_1_1BrokerFactory.html',1,'ace_time::internal']]],
-  ['brokerfactory_3c_20zoneinfo_2c_20zoneera_2c_20zonepolicy_2c_20zonerule_20_3e_326',['BrokerFactory&lt; ZoneInfo, ZoneEra, ZonePolicy, ZoneRule &gt;',['../classace__time_1_1internal_1_1BrokerFactory.html',1,'ace_time::internal']]]
+  ['basiczone_318',['BasicZone',['../classace__time_1_1BasicZone.html',1,'ace_time']]],
+  ['basiczonemanager_319',['BasicZoneManager',['../classace__time_1_1BasicZoneManager.html',1,'ace_time']]],
+  ['basiczoneprocessor_320',['BasicZoneProcessor',['../classace__time_1_1BasicZoneProcessor.html',1,'ace_time']]],
+  ['basiczoneprocessorcache_321',['BasicZoneProcessorCache',['../classace__time_1_1BasicZoneProcessorCache.html',1,'ace_time']]],
+  ['basiczoneprocessortemplate_322',['BasicZoneProcessorTemplate',['../classace__time_1_1BasicZoneProcessorTemplate.html',1,'ace_time']]],
+  ['basiczoneprocessortemplate_3c_20basic_3a_3abrokerfactory_2c_20basic_3a_3azoneinfobroker_2c_20basic_3a_3azoneerabroker_2c_20basic_3a_3azonepolicybroker_2c_20basic_3a_3azonerulebroker_20_3e_323',['BasicZoneProcessorTemplate&lt; basic::BrokerFactory, basic::ZoneInfoBroker, basic::ZoneEraBroker, basic::ZonePolicyBroker, basic::ZoneRuleBroker &gt;',['../classace__time_1_1BasicZoneProcessorTemplate.html',1,'ace_time']]],
+  ['brokerfactory_324',['BrokerFactory',['../classace__time_1_1internal_1_1BrokerFactory.html',1,'ace_time::internal']]],
+  ['brokerfactory_3c_20zoneinfo_2c_20zoneera_2c_20zonepolicy_2c_20zonerule_20_3e_325',['BrokerFactory&lt; ZoneInfo, ZoneEra, ZonePolicy, ZoneRule &gt;',['../classace__time_1_1internal_1_1BrokerFactory.html',1,'ace_time::internal']]]
 ];