+++ docs/html/search/files_1.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['common_2eh_379',['common.h',['../common_8h.html',1,'']]],
-  ['compat_2eh_380',['compat.h',['../compat_8h.html',1,'']]]
+  ['common_2eh_378',['common.h',['../common_8h.html',1,'']]],
+  ['compat_2eh_379',['compat.h',['../compat_8h.html',1,'']]]
 ];