+++ docs/html/search/variables_7.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['lastdeltaminutes_619',['lastDeltaMinutes',['../structace__time_1_1extended_1_1MatchingEraTemplate.html#a54213b73ec7e1341d4af32147566713b',1,'ace_time::extended::MatchingEraTemplate']]],
-  ['lastoffsetminutes_620',['lastOffsetMinutes',['../structace__time_1_1extended_1_1MatchingEraTemplate.html#ad6e089ef39f3531fae3c27705f1fc0aa',1,'ace_time::extended::MatchingEraTemplate']]],
-  ['letter_621',['letter',['../structace__time_1_1extended_1_1TransitionTemplate.html#a62e9a52a0d0d11548a80a44097b88fe6',1,'ace_time::extended::TransitionTemplate']]],
-  ['letters_622',['letters',['../structace__time_1_1internal_1_1ZoneContext.html#af0ea1d5c2d0615680a716125058e341f',1,'ace_time::internal::ZoneContext']]]
+  ['lastdeltaminutes_615',['lastDeltaMinutes',['../structace__time_1_1extended_1_1MatchingEraTemplate.html#a54213b73ec7e1341d4af32147566713b',1,'ace_time::extended::MatchingEraTemplate']]],
+  ['lastoffsetminutes_616',['lastOffsetMinutes',['../structace__time_1_1extended_1_1MatchingEraTemplate.html#ad6e089ef39f3531fae3c27705f1fc0aa',1,'ace_time::extended::MatchingEraTemplate']]],
+  ['letter_617',['letter',['../structace__time_1_1extended_1_1TransitionTemplate.html#a62e9a52a0d0d11548a80a44097b88fe6',1,'ace_time::extended::TransitionTemplate']]],
+  ['letters_618',['letters',['../structace__time_1_1internal_1_1ZoneContext.html#af0ea1d5c2d0615680a716125058e341f',1,'ace_time::internal::ZoneContext']]]
 ];