+++ src/ace_time/ZoneManager.h
@@ -184,14 +184,6 @@ class ZoneManagerTemplate : public ZoneManager {
       return Z(zoneInfo);
     }
 
-    /**
-     * Reset the transition cache of the zone processors in the cache.
-     * Useful when Epoch::currentEpochYear() is changed at runtime.
-     */
-    void resetZoneProcessors() {
-      mZoneProcessorCache.resetZoneProcessors();
-    }
-
   protected:
     /**
      * Constructor.